requirements.txt
kagglehub
openai>=1.0.0
pandas
python-dotenv
pillow
